--- src/main/java/Command.java
@@ -1,5 +1,0 @@
-interface Command {
-    public void execute(Storage storage, TaskList tasks, Ui ui) throws DownyException;
-    public boolean isExit();
-
-}
