+++ src/main/java/commands/ByeCommand.java
@@ -1,3 +1,9 @@
+package commands;
+
+import storage.Storage;
+import tasks.TaskList;
+import ui.Ui;
+
 public class ByeCommand implements Command {
     @Override
     public void execute(Storage storage, TaskList tasks, Ui ui) {
