+++ src/main/java/commands/DeadlineCommand.java
@@ -1,3 +1,11 @@
+package commands;
+
+import exceptions.DownyException;
+import storage.Storage;
+import tasks.Deadline;
+import tasks.TaskList;
+import ui.Ui;
+
 import java.time.LocalDateTime;
 
 public class DeadlineCommand implements Command {
