+++ src/main/java/commands/EventCommand.java
@@ -1,3 +1,11 @@
+package commands;
+
+import exceptions.DownyException;
+import storage.Storage;
+import tasks.Event;
+import tasks.TaskList;
+import ui.Ui;
+
 import java.time.LocalDateTime;
 
 public class EventCommand implements Command {
