+++ src/main/java/commands/HelpCommand.java
@@ -1,3 +1,9 @@
+package commands;
+
+import storage.Storage;
+import tasks.TaskList;
+import ui.Ui;
+
 public class HelpCommand implements Command {
 
     public void execute(Storage storage, TaskList tasks, Ui ui) {
