+++ src/main/java/commands/MarkCommand.java
@@ -1,3 +1,12 @@
+package commands;
+
+import exceptions.DownyException;
+import exceptions.InvalidFormatException;
+import storage.Storage;
+import tasks.Task;
+import tasks.TaskList;
+import ui.Ui;
+
 public class MarkCommand implements Command {
 
     public final String taskNumber;
@@ -12,9 +21,9 @@ public class MarkCommand implements Command {
             storage.markComplete(t);
             ui.displayCompletedTask(t);
         } catch (NumberFormatException e) {
-            throw new InvalidFormatException("Task number has to be a positive integer.");
+            throw new InvalidFormatException("tasks.Task number has to be a positive integer.");
         } catch (IndexOutOfBoundsException e) {
-            throw new IllegalArgumentException("Task number does not exist.");
+            throw new IllegalArgumentException("tasks.Task number does not exist.");
         }
     }
     public boolean isExit() {
