+++ src/main/java/exceptions/DownyException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class DownyException extends Exception {
 
     public DownyException(String message) { super(message); }
