+++ src/main/java/exceptions/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class InvalidCommandException extends DownyException {
     public InvalidCommandException() {
         super("Invalid command entered");
