+++ src/main/java/exceptions/InvalidFormatException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class InvalidFormatException extends DownyException {
     public InvalidFormatException(String message) {
         super(message);
