+++ src/main/java/exceptions/MissingArgumentException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class MissingArgumentException extends DownyException {
     public MissingArgumentException(String message) {
         super(message);
