+++ src/main/java/tasks/Deadline.java
@@ -1,7 +1,9 @@
-import java.time.LocalDateTime;
-import java.util.Date;
+package tasks;
 
-public class Deadline extends Task{
+import java.time.LocalDateTime;
+import parser.DateTimeHandler;
+
+public class Deadline extends Task {
     private final LocalDateTime time;
     public Deadline(String name, LocalDateTime time) {
         super(name);
